modules/locations-rest/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
flask==3.1.0
flask-accepts==1.0.1
flask-cors==5.0.0
flask-restx==1.3.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
markupsafe==3.0.2
marshmallow==3.23.1
packaging==24.2
pytz==2024.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
werkzeug==3.1.3
